add: middleware checkrole, seeder & code refactor
