+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 .env
+.env-prod
+.env-dev
 prisma/migrations
